[refactor] refactor zero and its test files
