+++ colossalai/gemini/__init__.py
@@ -1,10 +1,6 @@
-from .chunk import TensorInfo, Chunk, TensorState
-from .chunk_mgr import ChunkManager
+from .chunk import TensorInfo, TensorState
 from .stateful_tensor_mgr import StatefulTensorMgr
 from .tensor_placement_policy import TensorPlacementPolicyFactory
 from .gemini_mgr import GeminiManager
 
-__all__ = [
-    'StatefulTensorMgr', 'TensorPlacementPolicyFactory', 'GeminiManager', 'ChunkManager', 'TensorInfo', 'Chunk',
-    'TensorState'
-]
+__all__ = ['StatefulTensorMgr', 'TensorPlacementPolicyFactory', 'GeminiManager', 'TensorInfo', 'TensorState']
