+++ tests/test_gemini/update/test_zero_ddp.py
@@ -12,12 +12,12 @@ from functools import partial
 from tests.test_tensor.common_utils import set_seed
 from tests.components_to_test.registry import non_distributed_component_funcs
 from colossalai.nn.parallel import ZeroDDP
-from colossalai.zero.zero_optimv2 import ZeroOptimizerV2
+from colossalai.zero import ZeroOptimizer
 from colossalai.testing import parameterize
 from colossalai.gemini.gemini_mgr import GeminiManager
 from tests.test_tensor.common_utils import debug_print
 
-from colossalai.gemini.update import search_chunk_configuration, ChunkManagerV2
+from colossalai.gemini.chunk import search_chunk_configuration, ChunkManager
 
 
 @parameterize('placement_policy', ['cuda', 'cpu', 'auto'])
@@ -38,7 +38,7 @@ def exam_state_dict(placement_policy, keep_gathered):
     config_dict = search_chunk_configuration(model, search_range_mb=1, search_interval_byte=100)
     config_dict[world_size]['chunk_size'] = 5000
     config_dict[world_size]['keep_gathered'] = keep_gathered
-    chunk_manager = ChunkManagerV2(config_dict)
+    chunk_manager = ChunkManager(config_dict)
     gemini_manager = GeminiManager(placement_policy, chunk_manager)
     model = ZeroDDP(model, gemini_manager, pin_memory=True)
     model.train()
@@ -76,12 +76,12 @@ def exam_load_state_dict(placement_policy, keep_gathered):
         init_device = torch.device('cpu')
     else:
         init_device = None
-    chunk_manager = ChunkManagerV2(config_dict, init_device=init_device)
+    chunk_manager = ChunkManager(config_dict, init_device=init_device)
     gemini_manager = GeminiManager(placement_policy, chunk_manager)
     model = ZeroDDP(model, gemini_manager, pin_memory=True)
 
     optimizer = torch.optim.Adam(model.parameters())
-    optim = ZeroOptimizerV2(optimizer, model)    # initialize the link between chunk16 and chunk32
+    optim = ZeroOptimizer(optimizer, model)    # initialize the link between chunk16 and chunk32
 
     torch_dict = torch_model.state_dict()
     model.load_state_dict(torch_dict, strict=False)
